Each test generates individual set of output files
Changed the way, the combined output will be generated. So each test will
generate its own set of output files. The name of the output files are generated
according to:

	  tmp.TESTNUMBER.*

So the tests if the correct output was generated can now avoid diff, but use the
md5sum of the output instead.